Phase 1 (reproducibility): deterministic reproduce_final.py -> 0.8234 from frozen OOFs
Commits the frozen base OOFs (combo_v2, hier_v6, orient) + meta so a clean clone
reproduces the winning Kaggle submission with no training. reproduce_final.py runs the
full final pipeline (blend -> isotonic -> Saerens test-prior correction beta=2.0 ->
orient L2-injection -> robust threshold) and asserts the exact winning class counts
(L2=314, L3=559). Verified: OOF 0.7873, writes sub_pc_b20.csv == public 0.8234.

+++ scripts/rocket_rich_v2.py
@@ -40,7 +40,7 @@ def build_channels(seq):
         dd = np.diff(a, axis=1); return np.concatenate([dd[:, :1], dd], axis=1)
     jerk = np.sqrt(d(mx)**2 + d(my)**2 + d(mz)**2)
     return np.stack([mx, my, mz, sx, sy, sz, mag_m, mag_s, pitch, roll, incl, jerk,
-                     d(pitch), d(roll)], axis=1).astype(np.float32)
+                     d(pitch), d(roll)], axis=1).astype(np.float64)   # MultiRocket numba needs f64
 
 
 Xtr = build_channels(seqtr); Xte = build_channels(seqte)
